Changing verifier invocation from backend. Verifier will now be invoked in the appserver process. Deployment process will throw deployment exception and abort the deployment in case any exception is thrown from verifier or if any verifier assertion is failed. Verifier will run in appserver mode for this invocation. It will allow sun appserver specific features during verification. Standalone verifier will have options to run in appserver mode or portability mode.
Added a new constructor and a verify api in verifier which is used by deployment time verification. Verifier will use the descriptor generated during the deployment process. The appserver classloader is used for verification in this mode which allows applications to use appserver specific features.

Ran all AS samples and Quick look tests.

Reviewed by: Sanjeeb Sahoo.