Issue Number: 6282181, 6281748
The default-domain.xml.template file had a resource-ref jdbc/Pointbase
pointing to nothing. This was missed in my previous checkin for cleaning
out Pointbase from the appserver. This was causing the GUI and CLI to break
while trying to access the resource pointed to by the resource-ref. Removing
this resource-ref.

Reviewed By: Aditya Gore, Binod P G

Tests run: PE/EE Ql. Also, manually ensured that the resource-ref is no longer
around when a new domain is created.