Reviewed by: Ken Saks and Binod Pg
Local provider now returns a copy of the object looked up. This way we have simulated what the remote provider did via RMI. The problem with looking up the PersistenceManagerFactory resource in the CMP SQE test will go away. This is a generic solution and will take care of all types of resources.

Tests successfully ran:
1> SQE CMP test in question (ejb/cmp/cmpcmrApp)
2> EJB dev tests
3> PE QL
4> CTS smoke tests