Issue number: N/A Reviewed by:  Prashanth Abbagani
add asadmin.bat.template to this directory
update asadmin.templte to the current version  (the asadmin template files will be re
moved from the tools directory)
modified build.xml to publish the template file to the appropriate directory.
Tested this change by building and assembling the admin-cli and tools modules.