Issue number:  NA Obtained from: NA Submitted by:  Kedar Mhaswade Reviewed by:   Kedar Mhaswade, Jane Young
Modified Files:
	commands/src/java/com/sun/enterprise/cli/commands/CreateMBeanCommand.java
	commands/src/java/com/sun/enterprise/cli/commands/S1ASCommand.java

Fixing bugs with custom mbean related commands.
Proper handling of ObjectName passed by the users.
Proper handling of a generic List that backend returns. It was incorrect to compare List.class.getName(), because
java.util.List is an interface and not a class.

Tested manually. No QL impact.