Issue number: NA Obtained from: NA Submitted by:  Kedar Mhaswade Reviewed by:   Ron Monzillo         server.policy
Making sure that server does not complain with ACViolation
while loading (dynamically) the MBeans that are copied as
fully qualified classes in <instance-root>/applications/mbeans.

Currently, a very wide grant block has been specified because of
the bug: 6235678. Even though the grant block is less restricted,
to load the MBeans from anywhere, the underlying Java ClassLoader
makes sure that the classes loaded are from a restricted URL. Just
because of some other bug, if the grant block is not relaxed, then
an AccessControlViolation results.

Discussed with Ron and he's approved this till 6235678 gets fixed.

Ran QL successfully.