Modified Files:  	appserv-commons/src/java/com/sun/enterprise/deployment/archivist/AbstractArchive.java  	appserv-commons/src/java/com/sun/enterprise/deployment/archivist/FileArchive.java  	appserv-commons/src/java/com/sun/enterprise/deployment/deploy/shared/Archive.java  	appserv-core/src/java/com/sun/enterprise/deployapi/ProgressObjectImpl.java  	appserv-core/src/java/com/sun/enterprise/deployment/backend/DeploymentStatus.java  	appserv-core/src/java/com/sun/enterprise/deployment/client/DeployAction.java  	appserv-core/src/java/com/sun/enterprise/deployment/client/DeploymentClientUtils.java  	appserv-core/src/java/com/sun/enterprise/deployment/phasing/J2EECPhase.java Removed Files:  	appserv-commons/src/java/com/sun/enterprise/deployment/util/DescriptorLoader.java  	appserv-core/src/java/com/sun/enterprise/deployment/client/DescriptorLoader.java  	appserv-core/src/java/com/sun/enterprise/webservice/ClientWsdlPublisher.java
Description:

Moved the logic that collects wsdl for file publishing from
client to server.  This will remove the last DOL dependency
on the client side.

The list of wsdl files to be published is constructed in
J2EECPhase and passed to the client via DeploymentStatus.

Note that all of the wsdl related files are downloaded
from the server in this implementation.  Previously,
only the final wsdl is downloaded.  The rest of the wsdl
content are copied out of the archive on the client side
directly to the wsdl-publish-location specified in the
deployment descriptor.

Considering it is a corner case to have multiple wsdl
files to publish, the performance degregation (from
downloading multiple files) should be minimum in normal
cases.  We could also consider implementing downloading
a jar of wsdl files instead of one file at a time if
this ever becomes a problem.

Reviewer(s): Jerome Dochez, Hong Zhang

Tests Ran: (on Linux)
1. PE QL
2. deployment dev tests (w/ known annotation failures)
3. PE CTS
4. appserv-tests/devtest/webservice/googleserver
   (for file publishing unit test)