Issue number:  6276809 Obtained from: QE Submitted by:  Kedar Mhaswade Reviewed by:   Jean-francois Arcand
While invoking ant, through maven, the only way to do it in our Build System
is to modify this ant command line. This will fix the problem with build id because
Terena sends in the build_id on the top-level maven command line.

The side-effect: all "invoke-ant" maven goals are going to be affected because of this. The effect is not negative one, just an extraneous property.