fixed bug in AMXImplBase and other classes; they were not finishing their initialization until postRegister(). Rearranged code by factoring out preRegister().  This was causing failures for code that listened for registration of AMX MBeans, and that accessed the MBean before its postRegister() method was called.
removed XTypes.JMS_RESOURCE_CONFIG, JMSResourceConfig because
this type does not exist in domain.xml; this was a mistake in
the 8.1 implementation.

removed XTypes.SYSTEM_PROPERITES_CONFIG, SystemPropertiesConfig because
this type does not exist in domain.xml; this was a mistake in
the 8.1 implementation.

added support for creating/removing ConnectorServiceConfig in
ConfigConfig.

fixed API for MBeanRegistrationListener to not allow exceptions
from mbeanRegistered() and mbeanUnregistered().


Issue number:
Obtained from:
Submitted by:  Lloyd Chambers
Reviewed by:   amx-dev