Fix for 6296095: CTS smoke is failling when enabling keep-alive mechanism.
The JMX/Admin/JSR88 are openning connections and re-use the connection indefinitely.
When keep-alive is enabled. the connection was closed and an error messages was appended to the response, which confuse the client. Now the connection is closed without any advices (Grizzly was trying to be smart ;-)).

- All CTS smokes passed on Linux and Solaris
- All QL passed on Linux
- All JSP/Servlet tcks passed on Linux
- All devtests/web passed on Linux

Monitored the keep-alive count and make sure the connection was closed, but the client was still able to execute requests.

Reviewed by: Jan Luehe