Fix for 6305279 ("JSP cache taglib ignores 'scope' attribute"):
Added 2 new listeners: CacheRequestListener (instance of
javax.servlet.ServletRequestListener) and CacheSessionListener
(instance of javax.servlet.http.HttpSessionListener), which create a
cache (for JSP tag body invocations) and store it as a request or
session attribute, respectively. The cache is active during the
lifetime of the request or session.

This is in addition to the already existing CacheContextListener,
which stores the cache as a context attribute (application scope).

With this, the supported cache scopes are "request", "session", and
"application" (the latter being the default).

Reviewed by: pwc-dev

Tests run: QL/PE