Issue number: Obtained from: Siva, Shreedhar Submitted by:  Kedar Mhaswade Reviewed by:
This is an absolutely new API.
Nobody is calling it yet.
No QL impact whatsoever.

No changes to existing API's. The requirement came up:
- with MQ cluster requiring a support to know the "buddies" of a particular MQ broker that is managed "LOCALLY" by the Appserver.
- with a Iiop Service requiring to know the Iiop service-end-points of "buddies" of a server instance (in a cluster).

The API degenerates to empty values when the server instance is NOT part of a cluster.

Specific API's are provided for both the above requirements.

Tested thoroughly in a cluster using unit tests.