Adds support for "offline" AMX, which allows the AMX DomainRoot, QueryMgr, BulkAccessMgr, etc and all the AMX MBeans in com.sun.appserv.management.config to load in an in-process MBeanServer by instantiating com.sun.appserv.management.OfflineConfigIniter.
Once loaded, the resulting AMX MBeans function exactly the same as
they do when they are running in the DAS, except that they do not
support creation of new config elements (because that currently
is implemented using com.sun.appserv MBeans, which are not available
in this offline mode).  The logical next step is to make the
AMX MBeans implement creation of new config elements using ConfigBeans,
rather than com.sun.appserv MBeans.

The main change required to support this feature was an overhaul
of the Attribute name mapping code.  This was necessary because
the Attribute names in com.sun.appserv are different from those
found in the ConfigBeans eg Attribute "foo-bar" in the com.sun.appserv
MBean and "FooBar" in the config bean.  Most Attributes can be auto-mapped
but some are completely different eg AMX "Name" versus ConfigBean "Id"
and com.sun.appserv MBean "id".  The previous mapping code assumed
one mapping from "Name" to "Id", but a 1:N mapping was not considered.
The new AMXAttributeNameMapper class not only automates all but these
special cases, requiring no developer activity to do so, but works
against either the com.sun.appserv MBean or the ConfigBean.  A number
of special-casesstill exist for which a name cannot be deduced
(eg "Name" => "Id") and these are handled via matchName().  The net
result is a greatly simplified mapping mechanism which is both faster
and more efficient, with less code, and a reduced development burden.
In many cases, the previous addCustomMappings() method could be
removed completely from various AMX Impl classes.

Debugging code was added for the above, but that code does not
run unless the AMX-DEBUG system property is set.

There is as yet no dotted name support; that is next on the to-do list.

Unit tests were revised to take into account testing against the DAS and
in "offline" mode.  Most of the unit tests were touched by adding
the getCapabilities() method.

Various minor bug fixes, generic type improvements, etc were made
as they were discovered in the course of implementing these changes.

All unit tests except a few known exceptions pass, both against the
DAS and against "offline" AMX MBeans.

The very, very special case of AuthRealmConfig works provided that
the file is a real file path, and not a ${com.sun.appserv.instanceroot}
path, which requires yet more special code to deal with (will be done
if needed).  However, it didn't work at all before, so this is a big
improvement!  Javadoc was also greatly improved for AuthRealmConfig.



Issue number:
Obtained from:
Submitted by:  Lloyd Chambers
Reviewed by:   amx-dev