Modified Files:     avk/src/tools/com/sun/enterprise/tools/verifier/VerificationHandler.java
Description:

This directory contains fix for the following bug.

http://bt2ws.central.sun.com/CrPrint?id=6314022

Note that the cause of the bug is due to verifier removes the application
directory indiscriminatively.  When using directory deployment, or running
verifier on a directory, the application directory gets removed at the
end of the verification process.

The fix is to make sure the verifier does not clean up if the target of
verification is already a directory.

Note that the fix is slightly different between 8.1_02 and 9.0 because
the way that the verifier is invoked was different.

Reviewers: Vikas, Abhijit

Tests Ran:
1. PE QL
2. the reproduceable test case documented in the bug