Fix for 6292672 ("SJSAS 9.0 PE doesn't support dns-lookup-enabled which is needed for using /w allow/denyRemoteHost"):
Fix addresses an inconsistency between PE and EE domain.xml,
in that when using PE, one first has to add the <http-protocol>
element in order to be able to manipulate any of its
attributes (in this case: dns-lookup-enabled), whereas in EE,
the <http-protocol> element is already provided by default.
With this fix, <http-protocol> will be provided "out-of-the-box"
in PE as well.

Reviewed by: kedar.mhaswade

Tests run:
- QL/PE
- EE startup
- asadmin set --user admin server.http-service.http-protocol.dns-lookup-enabled=[true/false]
  now works "out-of-the-box" in PE