Issue number: bugs# 6308478 , 6308480 and 6308459 Reviewed by:  Prashanth Abbagani
Remove password and interactive as options in change-admin-password since password is prompted a
nd interactive is always true so no need for these options.
Add the "asadmin" as the group to pass in to the backend.
Remove hardcoded admin as the admin user, instead use the user provided by the command line for
--user option.
Modify ChangeAdminPasswordCommand to inherit from GenericCommand class instead of FileUserComman
d since FileUserCommand uses a different option for admin password.
Modify ChangeAdminPasswordCommand to prompt for the old/new password.
Add a polymorphic method called getPassword() so that warning message for password is not displa
yed.

Manually tested and passed QL tests.