Fix for 6319721 ("Remove BatchedReconfigSupport feature from PluggableFeatureFactory").
This feature, which allowed several changes to the server
configuration to be applied dynamically "in one go", was supported by
webcore (EE) only and is no longer needed in AS 9.0.

In AdminService.java, replaced invocation of
BatchedReconfigSupport.onStartup() with corresponding implementation,
which in the case of PE registers a WebModuleDeployEventListener.

Tests run:
- QL/PE