Fix for 6283880: [Grizzly] Sharing EE web configuration with PE
Setting acceptor-threads is now following the same behaviour as 8.1EE:

http://docs.sun.com/source/819-0082/hman1/create-http-listener.1.html

The number of acceptor threads for the listen socket. The recommended value is the number of processors in the machine. The default value is 1.

Grizzly map that value to the number of Selector it will creates to process requests.

- ALL QL passed when setting this value higher that 1 on PE

Approved by: CCC team
Reviewed by: Jan Luehe