Issue number:  6316801, 6317402 Obtained from: SQE Submitted by:  Kedar Mhaswade Reviewed by:   Byron Nevins Jr.
------
This is an unfortunate side-effect of an earlier checkin.
The issue is: We must have a server-identifier property while a custom mbean is registered in
the MBeanServer. But the MBean definition must not have this property in the configuration.

Thus an MBean whose ObjectName as far as config is considered is: "user:name=foo" will have
a runtime ObjectName of "user:name=foo,server=<target-server-id>".


No QL impact.
Ran the devtests for custom mbeans.
Also tested manually.