Bug 6316647 cannot list/create JMS physical destinations
Create and List are working fine based on a previous fix for bug# 6311390

Noticed problem with Delete and Purge because of the way the destination name is returned as part of detsination object name which is enclosed by double-quotes. Based on Siva's suggestion, exploring with IMQ to see if the double-quoting can be avoided in afuture release of MQ.

The fix is to remove the enclosed double quotes and the code works fine.

Performed pe/ee quick look tests.

Reviewed by: Siva

Files modified:
admin/mbeans/src/java/com/sun/enterprise/admin/mbeans/JMSDestination.java