Issue number: bugs#  6321608, 6317811, 6320200 Reviewed by:  bug# 6320200 is reviewed by Prashanth Abbagani               bugs# 6321608, 6317811 reviewed by self
For bug #6321608:
I modified the code in GenerateReportCommand.java to call displayExceptionMessage()
so the root cause of the exception message is displayed.

For bug# 6317811:
Backend has changed the type of bugids to String instead of String[] in CLIOptions class.
I modified the code in GenerateReportCommand.java to accomodate this change.

For bug# 63020200:
Check the length of strToEscape before checking for the next character in Command.java.


Manually tested the changes and passed QL tests.