Issue number:  6317208 Obtained from: Self Submitted by:  Kedar Reviewed by:   Jane/Kedar/Nandini
This is final checkin for this bug.
Verified that service name is used from the command line using --name option.
Verified that startinstances flag is honored while creating NA as a service.
Verified that PASSWORD_FILE contains AS_ADMIN_USER, PASSWORD and MASTERPASSWORD tokens.
Verified that messages are proper.
Verified that all the paths are absolute.

Made sure that domains and node-agents start after a reboot!

No QL impact.