Fix for 6311739 ("AS90EE:server_access_log file not used after the log rotation")
This also avoids redundant rotations if the datestamp suffix is not
changing.
For example, in PE, our default datestamp suffix contains just the
date (no hour/minutes/seconds), e.g., it looks something like:

  2005-09-14

Therefore, even if the rotation interval is set to 30s, there is no
need to rotate because the datestamp will not change for the rest of
09/14.
In order for rotation to occur in this case, a more fine-grained date
stamp pattern (that would include seconds) must be configured.

Reviewed by: Jeanfrancois Arcand

Tests run:
- QL
- Manual EE tests