Issue number: 6319375 Obtained from: Submitted by:  Shalini Muthukrishnan Reviewed by:   Sankara Rao
    Changes relating to the String monitor type and Constants for the 3 monitor types have been made. It is very hard in GUI to specialize the
case that it only depends on the existence of a property instead os its value. It is more effort to explain to user, when notifymatch and notifydiffer exist and which takes precedence. The following changes have thus been made. A property "stringnotify" has been added whose value would be either "notifymatch" or "notifydiffer". "notifymatch" would have higher priority than "notifydiffer".