Issue number:  6325922 Obtained from: Sankar Submitted by:  Kedar Mhaswade Reviewed by:   Self
Made sure that the startinstances flag on creating a node agent
as a service is "true" by default. If the administrator does not want the server instances to be started on the node agent startup, then s/he needs to specify "--serviceproperties startinstances=false" on create-service command line.

Tested manually with both the values.
No QL impact.

This brings the create-service default with start-node-agent default w.r.t. this flag, startinstances.