Comments: Warn if app client does not specify Main-Class in its manifest
Bug Number and bug synopsis:
Comments: Add non-fatal warning if appclient does not specify Main-Class in manifest

Bug Number and bug synopsis: 6318043 - Behavior of -mainclass option to appclient should change for Java EE 5.0 appclients.

Problem: Java EE 5 allows components (including app clients) to omit descriptors and use annotations.  The example in the bug used annotations but did not specify the Main-Class attribute in its manifest.  The annotation processing of app clients during deployment uses the Main-Class to select the class to inspect to find annotations.  With Main-Class not specified, deployment could not search for annotations and could not prepare injections for runtime.

At runtime, the client triggered a null pointer exception when it tried to refer to a field that it expected to be injected.

The spec seems to require that app clients specify Main-Class in their manifests.  We have not enforced this in earlier versions of the app server, and we did not want to cause apps that might have deployed earlier to start failing.  Yet omitting Main-Class is an important oversight on the part of the developer and we wanted to flag this.  The changes below add logic that issues a warning to server.log and as a substage DeploymentStatus that is reported back to the deploying client if an app client does not specify Main-Class in its manifest.


Reviewed by: Hong


Tests run: SQE quicklook PE and EE tests; deployment PE devtests; manual test of an ear containing a client that does not specify Main-Class

Files affected: all in appserv-core/src/java/com/sun/enterprise/deployment/backend

Deployer.java
AppDeployer.java
AppClientModuleDeployer.java
LocalStrings.properties


6327786 - Failure loading ear during restart if embedded app client has no application-client.xml file

Problem:

An app client embedded in an ear had no developer-provided descriptor and used no annotations.  The annotation processing logic does not generate a descriptor in this case.  During app server restart, such ears failed to load because no descriptor information was made available.

At Hong's suggestion, when providing an app's descriptor if no descriptor info is present either from the developer or as a result of annotation processing then the archivist should provide default DD info in all cases.

Reviewed by: Hong


Tests run: SQE quicklook PE and EE tests; deployment PE devtets; manual test of ejb30 deployment dev test with annotations across server restart

Files affected: in appserv-commons/src/java/com/sun/enterprise/deployment/archivist/

Archivist.java
LocalStrings.properties (removed a message that is no longer used - it reported a missing descriptor which is OK now in Java EE 5)