Fix for 6330300 ("URL forwarding feature as in AS7.x is gone from AS8.x"):
Added support for "redirect" property (on virtual-server), whose
syntax conforms to that of the "redirect" directive in the webserver's
obj.conf.

"redirect" properties are parsed and stored in the
VirtualServerPipeline (see 6331392: "Preparation work for 6330300") of
the associated virtual server. VirtualServerPipeline inspects incoming
request URI and tries to match it against the "from" URI prefix of its
"redirect" properties. The longest match wins.

If a match has been found, a redirect is issued to the target URL
computed from the property's "url" or "url-prefix" elements,
respectively.

Otherwise, the VirtualSeverPipeline executes its valves.

Reviewed by: Jeanfrancois Arcand

Tests run:
- QL/PE