Partial fix for 6304161: 9.0EE:Logout button reloads admingui, when GUI is accessed with http port, after disabling Security.
When stopping the admin-listener, make sure all pending connections are properly closed before. This will avoid the admin-gui looping for a response. Note that there will be no response because the socket will be closed during the same transaction of reconfiguring the admin-listener.

The solution for admin/admin-gui is to flush the response *before* reconfiguring the admin-listener (with the disavange of loosing a failure if the reconfig doesn't work).

- All QL passed
- Manualy tested the admin-gui

Reviewed by: Jan Luehe