Issue number: 6287047 and 6331540 Reviewed by:  Prashanth Abbagani
I know 6331540 is a P3, but by fixing 6287047, it also fixes this bug.
Anyway, I'm putting the jar files back in the manifest file except for appserv-ws.jar.
 The reason for doing this is that some testscripts and samples are directly calling th
e classes from the jar files not using sun-appserv-admin task which invokes AppservClas
sLoader.java.  Therefore asant was broken when they are trying to directly invoke some
classes from appserver jar files.

The reason why appserv-ws.jar is removed is because of bug 6325258.  appserv-ws.jar con
tain it's own ant task.  When this jar file is added to the manifest file or the classp
ath it will get called first which caused a problem when the jaxws-test script is tryin
g to define it's own ws tasks that uses the same name in appserv-ws.jar.

tested on both Win and Unix build verified that jaxws-test scripts works with asant and
 also did some simple tests with asant to make sure that it's not broken.