Issue number:  6340373 Reviewed by:   Dinesh.Patil@sun.com
IN=81052 - Remove dbschema.jar from appserver's lib directory

The classes contained in the dbschema.jar library are used only by the cmp
component of the appserver. These classes are also packaged into
appserv-cmp.jar, which represents the cmp-component. Issues:

1. dbschema classes are duplicated in the appserver's classpath: Once in
$AS_HOME/lib/dbschema.jar and also in $AS_HOME/lib/appserv-cmp.jar

2. As dbschema classes are used only by cmp, they should not appear directly in
$AS_HOME/lib

3. Repackaging dbschema classes into appserv-cmp.jar reduces the number of jar
files in the classpath, and provides a performance benefit according to the
performance team

After this check in, users must call "maven bootstrap-all" from
glassfish/bootstrap in order to build the cmp component. This will move
${glassfish.home}/lib/dbschema.jar to
${glassfish.home}/../DBSCHEMA/dbschema.jar. The property ${dbschema.jar} is
changed to point to the new dbschema location.