Fix for 6345104: Bigapps(Nilesapp) stress test on appserver 9.0 build returns keepalive exceptions
On win32, the NIO Selector get easily overloaded, and the use of an alternate Selector is required to avoid a performance bottleneck. The stress test (on win32) overload the Selector and produce a thread race where the counter used to track the alternate Selector get updated twice, producing the exception. Use a synchronized Factory instead to dispatch Selector amongs client.

- All QL passed on win32
- All devtests/web passed on win32

Reviewed/developped with Scott Oaks.