Fix for bug 6345626: The images of a war that are being served by the web server are being locked so you can't undeploy
On win32, resources are locked even if you close the FileChannel associated with a MappedByteBuffer. The only way to remove the lock is to remove the reference so the gc can garbage the MappedByteBuffer. Thus, when undeploying, all the cache entries related to a WebModule must be removed and MappedByteBuffer de-referenced.

- All QL passed on win32 and Linux
- Manually tested the problem on win32 by deploying an app, invoking a static resource and then undeploying the app.

Reviewed by: Jan Luehe and Tim Quinn