Comments: Fix 6344079 - Java Web Start support throws NPE if module does not contain an accessible class
In an ear, a submodule was mislabeled in application.xml as an app client (it was in a <java> element) instead of an EJB module. The app server treated it as an app client, including broadcasting application events as it was loaded. Therefore the Java Web Start feature tried to treat it as an app client also and could not. This led to the NPE.

This error has been seen only once, and that was in this situation when an EJB submodule was incorrectly identified as an app client in the application.xml. Still, this is a low-cost runtime check that can prevent similar problems in the future.

Reviewed by: Qingqing

Tests run: SQE quicklook PE tests; deployment PE devtests; CTS smoke tests; manual test using an ear attached to the CR

Files affected:
glassfish/appserv-core/src/java/com/sun/enterprise/appclient/jws
  AppclientJWSSupportInfo.java
  LocalStrings.properties