Issue number:  6239630 Reviewed by:   Rochelle.Raccah@sun.com IN=83030 - Allow disabling ModelValidation check for FK field in DFG
This check in is essentially a backport for the priliminary fix of CR 6239630.
To incease the efficiency for SPECj a foreign key field needs to be mapped
explicitly to the default fetch group (DFG). The fix introduces the flag
'AllowManagedFieldsInDefaultFetchGroup' that disables a ModelValidation check.
This flag must only be used if the relationship is read only, i.e. never changed
by the application.