Issue number:  6286325 Obtained from: QE Submitted by:  Lidia Reviewed by:   Self
Basically, this is an issue with a "correct" access-log element not showing up in the domain.xml out of the box.
This hampers automated tests with CLI unless a GUI operation is done to set the values on access-log of a virtual
server.

This check-in makes sure that domain gets a properly configured access-log element in PE.
The values of access-log are different from the defaults in domain.xml's dtd because these
values make more sense for PE.

I have discussed this with Jan Luehe till he was bored :)

Made sure that access-log is in proper format, server likes it and it (server) starts/stops properly.