Issue number:  6196993 Obtained from: Submitted by:  bnevins Reviewed by:   kedar mhaswade
summary:  Say you backup "domain1".  You are not allowed to restore the domain to a differently named domain.
Before the changes, restore-domain allowed it which resulted in a domain that wouldn't start.
After the changes, an error is issued and it refuses to do the restore.
Here is the actual error message:

asadmin> restore-domain --filename C:/ee/domains/domain1/backups/sjsas_backup_v00001.zip qbert
com.sun.enterprise.config.backup.BackupException: A domain can only be restored to the original domain name.  The original backed-up domain
name is domain1, and you are attempting to restore it to qbert.
asadmin>

-------------------------------

It was pretty easy -- most of the code I needed was already written!  I did this:

1) Add an error message to localstrings
2) check the domain-name for the restore against the domain name that is stored inside the backup zips "status.properties" file
3) if the names aren't the same -- error out.
4) convenience method (read(File) which calls read(File f, true)) was added to StatusFile.java