Issue number:  6331988 Reviewed by:   MiteshMeswani@Sun.Com IN=83422 - Add special Oracle support as precompiled library for cmp
CMP's special Oracle support is excluded from the build, as the Oracle JDBC
driver must be present for compilation. A pre-compiled version will checked in
under <cmp-basedir>/release/build/OracleSpecialDBOperation.jar. This archive is
unpacked for every build, and must be updated on every change to this the class
OracleSpecialDBOperation.java