Incremental fix for bug 6357309: Reduce GC impact in web container
The WebServices problem is perhaps a little unrealistic; their test
creates thousands of very short-lived sessions. For each session, the session object gets created. With a Hashtable, the size of the session is 96 bytes (not including what gets stored in there -- that's
just the hashtable). The size of a HashMap is 120 byes, but the size of a ConcurrentHashMap is 1272 bytes. So we end up being gc-throttled in the test.

- All QL passed.

Reviewed by: Scott Oaks.