Issue number:  entity-persistence-tests module Obtained from: TopLink Submitted by:  Tom Ware Reviewed by:   TopLink Team
This is the intial testing check-in for the entity-persistence module. This module does not participate in the general GlassFish build and should be built independantly.

The readme.txt file at the base directory explains how to run these tests.

Tested by building entity-persistence, running the maven targets in entity-persistence-testing and running the tests being checked in.