Fix for bug 6342889
Reviewed by : Ken Saks, Sankara Rao.

Description.
-----------
Application's invocation chain is stored as an InheritableThreadLocal.
So, whenever a thread is created it inherits the invocation chain of the
parent thread. This is for handling threads created from servlets.

It so happened that, when on-demand initialization is on, ORB thread pool
got created from within the admin servlet. And it inherited the servlet's
invocation chain. That screwed up the SecurityMechanismSelector logic.

The fix is to introduce the custom invocation type for servicegroup startup,
so that any invocation of the parent thread is not used, during the startup
flow.

Tests: PE QL, cts-smoke, on-demand devtests. EE QL