Incremental fix for 4965017 ("polling of expired sessions causing high cpu utilization"):
Instead of commenting out the dynamic reload facility of a
StandardContext (which is not needed in AS or WS, because both AS and
WS have their own reload mechanism), it can now be enabled or disabled
programmatically by configuring a new "reload" property on StandardContext.
By default, "reload" is set to false.

This will be useful if PWC is embedded in hosting applications other than
AS or WS, which may not have their own reload mechanism and therefore need
to leverage the StandardContext's reload feature.

Reviewed by: Jeanfrancois Arcand