Issue number:  94 - Need to get PRE_UNDEPLOY callback from undeployment of a war Reviewed by:   Hong Zhang Passed QL, deployment dev tests.
Files Modified:
In appserv-core module:
com/sun/enterprise/deployment/backend/ModuleDeployer.java

For a war module read the deployed directory and get all the persistence unit descriptors that might be defined for that module. We need this information to ensure that all the tables that were created as part of deployment are successfully dropped from the database.