Fixes for:
6360005 - Code invoking getResourceAsStream can lock jar files on Windows
6349295 - Cannot run .ear application when installation is in directory with space on Window

Re: 6360005 - When an app invokes getResourceAsStream, the JDK's default stream handler for a jar URL causes a connection to the jar file to be cached in a static data structure and, therefore, the jar file becomes locked.  Once cached, the jar file cannot be overwritten, renamed, or deleted.  This makes redeployment or undeployment of the app impossible without restarting the server.  The changes in EJBClassLoader intercept invocations of the getResourceAsStream method and prevent the JDK's caching (since the EJBClassLoader is already caching connections to the jars anyway).

Re: 63492925 - When the app server is installed or bootstrapped on Windows into a path that contains a space, there were several problems.  Most had to do with using the construct someFile.toURL() to pass to the EJBClassLoader's appendURL(URL url) method.  The File.toURL method does not encode special characters such as spaces, so the resulting URL is illegal if the path does contain a space.  There were changes to EJBClassLoader, primarily to add a new method, appendURL(File file).  This allows the invoking code to pass the file and the logic in appendURL makes sure to build the URL correctly.

Most of the changes to other classes are very small, using URIs (which provide the proper encoding) or using the newer appendURL method.


Reviewed by: Qingqing (additional feedback pending from Siva, Sahoo)

Tests run:
Windows with no blank in path: SQE PE quicklook tests and CTS smoke tests; manual verifier and deploy --verify=true tests

Windows with blank in path:    SQE PE quicklook tests - all passed except web service tests which failed due to improper encoding of message security config file path.  This problem predates these check-ins and will be addressed in a subsequent check-in.  manual verifier and deploy --verify=true tests

Linux (thanks, Qingqing):      SQE PE quicklook tests and CTS smoke tests



Files affected:
  appserv-commons/src/java/com/sun/enterprise/deployment/node/SaxParserHandler.java
  appserv-commons/src/java/com/sun/enterprise/deployment/util/ModuleContentLinker.java
  appserv-commons/src/java/com/sun/enterprise/util/FileUtil.java
  appserv-commons/src/java/com/sun/logging/enterprise/system/core/classloading/LogStrings.properties
  appserv-commons/src/java/com/sun/logging/enterprise/system/util/LogStrings.properties
  appserv-commons/src/java/org/apache/tools/ant/taskdefs/optional/sun/appserv/AppservClassLoader.java

  appserv-core/src/java/com/sun/enterprise/appclient/Main.java
  appserv-core/src/java/com/sun/enterprise/connectors/util/RARUtils.java
  appserv-core/src/java/com/sun/enterprise/connectors/util/ConnectorClassLoaderUtils.java
  appserv-core/src/java/com/sun/enterprise/deployapi/DeploymentFactoryInstaller.java
  appserv-core/src/java/com/sun/enterprise/deployment/backend/DeploymentUtils.java
  appserv-core/src/java/com/sun/enterprise/instance/EJBModulesManager.java
  appserv-core/src/java/com/sun/enterprise/loader/ClassLoaderUtils.java
  appserv-core/src/java/com/sun/enterprise/loader/EJBClassLoader.java
  appserv-core/src/java/com/sun/enterprise/util/JarClassLoader.java
  appserv-core/src/java/com/sun/enterprise/webservice/WsUtil.java
  appserv-core/src/java/com/sun/enterprise/webservice/codegen/JaxRpcRICodegen.java

  avk/src/tools/com/sun/enterprise/tools/verifier/VerificationHandler.java
  avk/src/tools/com/sun/enterprise/tools/verifier/tests/web/WebTestsUtil.java
  avk/src/tools/com/sun/enterprise/tools/verifier/web/WebVerifier.java